Fix failing tests in class_method_test
I was seeing 4 failures when running the tests in this test case:

1. ClassMethodTest#test_should_call_remove_new_method: NoMethodError:
   undefined method `reset_mocha' for #<Class:0x007fa3e00a9b38>

2. ClassMethodTest#test_should_call_restore_original_method:
   NoMethodError: undefined method `reset_mocha' for
   #<Class:0x007fa3e00a8328>

3. ClassMethodTest#test_should_restore_original_method_accepting_a_block_parameter:
   NoMethodError: undefined method `method_x' for #<Class:0x007fa3e009bb50>

4. ClassMethodTest#test_should_restore_original_method
   [test/unit/class_method_test.rb:69]: Expected false to be truthy.

The failures are all related to the fact that methods defined in
`Mocha::ObjectMethods` weren't available in the test.

These tests were passing when run as part of `rake test` because
`Mocha::ObjectMethods` was being included elsewhere in the test suite.

I ran this test case and `rake test` in Ruby versions 1.8.7, 1.9.3,
2.0.0, 2.1, 2.2 and 2.3 after making this change. There were no
failures.