Fix failing tests in any_instance_method_test
I was seeing 2 failures when running the tests in this test case:

1. AnyInstanceMethodTest#test_should_call_remove_new_method:
   Mocha::ExpectationError: unexpected invocation:
   #<Mock:0x7fe5d756b958>.reset_mocha()

2. AnyInstanceMethodTest#test_should_call_restore_original_method:
   Mocha::ExpectationError: unexpected invocation:
   #<Mock:0x7fe5d756b840>.reset_mocha()

The failures are all related to the fact that methods defined in
`Mocha::ObjectMethods` weren't available in the test.

These tests were passing when run as part of `rake test` because
`Mocha::ObjectMethods` was being included elsewhere in the test suite.

I ran this test case and `rake test` in Ruby versions 1.8.7, 1.9.3,
2.0.0, 2.1, 2.2 and 2.3 after making this change. There were no
failures.